chore(home): Moved around components, divided 3D Models from others
