+++ src/components/3d_models/NonEuclideanCube.tsx
